Refactor the menu rendering options
