+++ decidim-core/app/presenters/decidim/menu_item_presenter.rb
@@ -28,7 +28,8 @@ module Decidim
     def render
       content_tag :li, class: link_wrapper_classes do
         output = [link_to(composed_label, url)]
-        output.push(@view.send(@menu_item.submenu).render) if @menu_item.submenu
+        output.push(@view.send(:simple_menu, @menu_item.submenu).render) if @menu_item.submenu
+
         safe_join(output)
       end
     end
