--- decidim-participatory_processes/app/helpers/decidim/participatory_processes/admin/participatory_processes_menu_helper.rb
@@ -1,17 +1,0 @@
-# frozen_string_literal: true
-
-module Decidim
-  module ParticipatoryProcesses
-    module Admin
-      module ParticipatoryProcessesMenuHelper
-        def admin_participatory_process_attachments_menu
-          @admin_participatory_process_attachments_menu ||= simple_menu(target_menu: :admin_participatory_process_attachments_menu)
-        end
-
-        def admin_participatory_process_components_menu
-          @admin_participatory_process_components_menu ||= simple_menu(target_menu: :admin_participatory_process_components_menu, options: { container_options: { id: "components-list" } })
-        end
-      end
-    end
-  end
-end
